test(coderd/x/chatd): consolidate model-call shape tests into flow owners
Fold standard-turn and compaction-summary template assertions into the
reasoning-effort clamp test, manual-title provider-option assertions
into the synthetic API key test, and status-label envelope assertions
into the structured status-label test. Replace the end-to-end summary
fixture with a focused resolver-level omission test that also covers
summaryObjectCall. Delete modelcall_shape_internal_test.go.
